Rename MainAppController
