+++ src/main.tsx
@@ -6,7 +6,7 @@
 import 'react-native-gesture-handler';
 
 import React from 'react';
-import { MainAppCtl } from './MainApp/MainAppCtl';
+import { MainAppCtl } from './MainApp/MainAppController';
 
 export const Main = () => {
   return <MainAppCtl />;
